+++ pages/accounts/settings/index.tsx
@@ -1,0 +1,14 @@
+import Navbar from "@/components/Navbar";
+import Head from "next/head";
+
+export default function SettingsPage() {
+  return (
+    <>
+      <Head>
+        <title>Settings | Poster</title>
+      </Head>
+      <div>Settings</div>
+      <Navbar />
+    </>
+  );
+}