+++ src/features/Profile/components/Header.tsx
@@ -1,8 +1,10 @@
+import { PageUrls } from "@/@enums/pages";
 import type { User } from "@/@types/user";
 import Avatar from "@/components/Avatar";
+import LinkButton from "@/components/LinkButton";
 import { useAccount } from "@/context/AccountContext";
 import styles from "../Profile.module.css";
-import ConnectionButton from "./ConnectionButton";
+import ConnectionButton from "./ConnectionStatus";
 
 export default function Header({
   id,
@@ -11,6 +13,7 @@ export default function Header({
   connectionStatus,
 }: User) {
   const { user } = useAccount();
+
   return (
     <div className={styles.header}>
       <Avatar className={styles.avatar} />
@@ -19,8 +22,12 @@ export default function Header({
           <div className={styles.fullname}>{fullname}</div>
           <div className={styles.username}>@{username}</div>
         </div>
-        {user?.id !== id && (
+        {user?.id !== id ? (
           <ConnectionButton connectionStatus={connectionStatus} />
+        ) : (
+          <LinkButton href={PageUrls.SETTINGS} className={styles.settings}>
+            Settings
+          </LinkButton>
         )}
       </div>
     </div>