+++ src/graphql/@types/auth.ts
@@ -1,0 +1,13 @@
+import type { User } from "@/@types/user";
+
+type AuthToken = {
+  token: string;
+  expiredAt: string;
+};
+
+export type LoginQuery = {
+  login: {
+    user: User;
+    authToken: AuthToken;
+  };
+};