+++ src/graphql/@types/post.ts
@@ -1,0 +1,7 @@
+import type { Post } from "@/@types/post";
+
+export type GetUserPostsByUsernameQuery = {
+  getUserPostsByUsername: {
+    posts: Post[];
+  };
+};